requirements.txt

requests==2.31.0
gunicorn         
aiohttp         
python-telegram-bot==20.7
flask